split up contour index
